feat: add reprocess command-line support and refactor services
- Add new processor: llmwhisperer_analytical.py
- Add services package and move gcp.py, llmwhisperer.py into it
- Update launch.json and constants.py for new structure
- Refactor and update analytical.py, main.py, and docling_analytical.py
- Move and update tests for llmwhisperer service
