+++ src/main.py
@@ -1,10 +1,20 @@
+import os
+
 import typer
 from dotenv import load_dotenv
 
+from analytical import reprocess as reprocess_analytical
 from analytical import run as run_analytical
-from processors.docling_analytical import process_file as process_file_docling
-from processors.llmwhisperer import process_file as processs_file_llmwhisperer
-from utils.constants import MethodType
+from processors.docling_analytical import (
+    process_pdf_file as process_pdf_file_docling,
+)
+from processors.llmwhisperer_analytical import (
+    process_txt_file as process_txt_file_llmwhisperer,
+)
+from services.llmwhisperer import (
+    process_pdf_file as process_pdf_file_llmwhisperer,
+)
+from utils.constants import FileType, MethodType
 from utils.spliter import split_pdf_to_pages
 
 app = typer.Typer()
@@ -19,11 +29,11 @@ app.add_typer(spliter_app, name="spliter", help="Splitting commands")
 )
 def run(
     path: str,
-    output_dir: str = "output",
+    output_dir: str = os.path.join(os.getcwd(), "output"),
     start: int = 1,
     end: int | None = None,
     upload: bool = False,
-    processed_dir: str = "processed",
+    processed_dir: str = os.path.join(os.getcwd(), "processed"),
     reprocess: bool = False,
     method: MethodType = MethodType.llmwhisperer,
 ):
@@ -35,9 +45,34 @@ def run(
         upload=upload,
         reprocess=reprocess,
         processed_dir=processed_dir,
-        process_file_fn=processs_file_llmwhisperer
+        process_pdf_file_fn=process_pdf_file_llmwhisperer
         if method == MethodType.llmwhisperer
-        else process_file_docling,
+        else process_pdf_file_docling,
+        process_txt_file_fn=process_txt_file_llmwhisperer
+        if method == MethodType.llmwhisperer
+        else None,
+    )
+
+
+@analytical_app.command(
+    help="Run the analytical extraction process on a PDF file with a different approach"
+)
+def reprocess(
+    path: str,
+    output_dir: str = os.path.join(os.getcwd(), "output"),
+    method: MethodType = MethodType.llmwhisperer,
+    file_type: FileType = FileType.TXT,
+):
+    return reprocess_analytical(
+        path,
+        output_dir,
+        process_pdf_file_fn=process_pdf_file_llmwhisperer
+        if method == MethodType.llmwhisperer
+        else process_pdf_file_docling,
+        process_txt_file_fn=process_txt_file_llmwhisperer
+        if method == MethodType.llmwhisperer
+        else None,
+        file_type=file_type,
     )
