+++ tests/services/test_llmwhisperer.py
@@ -2,10 +2,10 @@ from unittest.mock import MagicMock, patch
 
 from unstract.llmwhisperer.client_v2 import LLMWhispererClientException
 
-from processors.llmwhisperer import process_file
+from services.llmwhisperer import process_pdf_file
 
 
-@patch("processors.llmwhisperer.LLMWhispererClientV2")
+@patch("services.llmwhisperer.LLMWhispererClientV2")
 def test_process_file_success(mock_client_cls, tmp_path):
     # Arrange
     mock_client = MagicMock()
@@ -17,7 +17,9 @@ def test_process_file_success(mock_client_cls, tmp_path):
     test_pdf.write_text("dummy pdf content")
 
     # Act
-    output_path = process_file(str(test_pdf))
+    output_path = process_pdf_file(
+        str(test_pdf), str(test_pdf).replace(".pdf", ".txt")
+    )
 
     # Assert
     assert output_path == str(test_pdf).replace(".pdf", ".txt")
@@ -25,7 +27,7 @@ def test_process_file_success(mock_client_cls, tmp_path):
         assert f.read() == "extracted text"
 
 
-@patch("processors.llmwhisperer.LLMWhispererClientV2")
+@patch("services.llmwhisperer.LLMWhispererClientV2")
 def test_process_file_exception(mock_client_cls, tmp_path):
     # Arrange
     mock_client = MagicMock()
@@ -34,6 +36,8 @@ def test_process_file_exception(mock_client_cls, tmp_path):
     test_pdf = tmp_path / "test.pdf"
     test_pdf.write_text("dummy pdf content")
 
-    output_path = process_file(str(test_pdf))
+    output_path = process_pdf_file(
+        str(test_pdf), str(test_pdf).replace(".pdf", ".txt")
+    )
     # Assert
     assert output_path == ""
